requirements/development.txt
-r base.txt
django-debug-toolbar>=1.3.0
Werkzeug>=0.10.1
gunicorn==19.3.0
